refactor: Receipt의 증정품 증정 관련 로직 캡슐화
