+++ docs/README.md
@@ -54,7 +54,8 @@
     
 ### 추가 구현 및 수정 사항  
 - [x] Promotion -> Event / Event -> Giving으로 네이밍 변경  
-- [ ] Receipt의 증정품 증정 관련 로직을 최대한 캡슐화해보기  
+- [x] Receipt의 증정품 증정 관련 로직을 최대한 캡슐화해보기  
+- [ ] 메소드 및 변수명 체크  
 - [ ] README에 프로젝트 소개 추가  
   
 ### 구현 세부 사항
